+++ packages/eyes-cypress/src/plugin/pluginExport.js
@@ -1,23 +1,23 @@
 'use strict';
+const setGlobalRunHooks = require('./hooks');
+const CYPRESS_SUPPORTED_VERSION = '6.2.0';
 
-function makePluginExport({startServer, config}) {
+function makePluginExport({startServer, eyesConfig, visualGridClient, logger}) {
   return function pluginExport(pluginModule) {
     let closeEyesServer;
     const pluginModuleExports = pluginModule.exports;
     pluginModule.exports = async (...args) => {
       const {eyesPort, closeServer} = await startServer();
-
       closeEyesServer = closeServer;
       const moduleExportsResult = await pluginModuleExports(...args);
-      const eyesConfig = {
-        eyesIsDisabled: !!config.isDisabled,
-        eyesBrowser: JSON.stringify(config.browser),
-        eyesLayoutBreakpoints: JSON.stringify(config.layoutBreakpoints),
-        eyesFailCypressOnDiff:
-          config.failCypressOnDiff === undefined ? true : !!config.failCypressOnDiff,
-        eyesDisableBrowserFetching: !!config.disableBrowserFetching,
-      };
-      return Object.assign(eyesConfig, {eyesPort}, moduleExportsResult);
+      const [on, config] = args;
+
+      if (config.version >= CYPRESS_SUPPORTED_VERSION && config.experimentalRunEvents) {
+        setGlobalRunHooks(on, {visualGridClient, logger});
+        eyesConfig.eyesLegacyHooks = false;
+      }
+
+      return Object.assign({}, eyesConfig, {eyesPort}, moduleExportsResult);
     };
     return function getCloseServer() {
       return closeEyesServer;