+++ packages/eyes-cypress/src/plugin/runningTests.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const runningTests = {
+  tests: [],
+  add(test) {
+    this.tests.push(test);
+  },
+  reset() {
+    this.tests = [];
+  },
+};
+
+module.exports = runningTests;