+++ packages/eyes-cypress/test/it/plugin/startPlugin.test.js
@@ -18,13 +18,13 @@ describe('start plugin', () => {
   });
 
   it('starts plugin server and patches module exports', async () => {
-    const {eyesPort} = await __module.exports();
+    const {eyesPort} = await __module.exports(() => {}, 'test');
     const resp = await fetch(`https://localhost:${eyesPort}/hb`);
     expect(resp.status).to.equal(200);
   });
 
   it('patches module exports with correct pref', async () => {
-    const {eyesIsDisabled, eyesFailCypressOnDiff} = await __module.exports();
+    const {eyesIsDisabled, eyesFailCypressOnDiff} = await __module.exports(() => {}, 'test');
     expect(eyesIsDisabled).to.be.false;
     expect(eyesFailCypressOnDiff).to.be.true;
   });
@@ -47,7 +47,7 @@ describe('start plugin', () => {
     });
 
     it('patches module exports with disabled eyes pref', async () => {
-      const {eyesIsDisabled} = await __module.exports();
+      const {eyesIsDisabled} = await __module.exports(() => {}, 'test');
       expect(eyesIsDisabled).to.be.true;
     });
   });
@@ -70,7 +70,7 @@ describe('start plugin', () => {
     });
 
     it('patches module exports with dont fail on diff pref', async () => {
-      const {eyesFailCypressOnDiff} = await __module.exports();
+      const {eyesFailCypressOnDiff} = await __module.exports(() => {}, 'test');
       expect(eyesFailCypressOnDiff).to.be.false;
     });
   });