+++ packages/eyes-cypress/test/unit/plugin/handlers.test.js
@@ -3,6 +3,7 @@ const {describe, it, beforeEach} = require('mocha');
 const {expect} = require('chai');
 const makeHandlers = require('../../../src/plugin/handlers');
 const {PollingStatus} = require('../../../src/plugin/pollingHandler');
+const runningTests = require('../../../src/plugin/runningTests');
 const {promisify: p} = require('util');
 const psetTimeout = p(setTimeout);
 
@@ -32,11 +33,8 @@ describe('handlers', () => {
   }
 
   beforeEach(() => {
-    handlers = makeHandlers({
-      makeVisualGridClient: () => ({
-        openEyes: fakeOpenEyes,
-      }),
-    });
+    const visualGridClient = {openEyes: fakeOpenEyes};
+    handlers = makeHandlers({visualGridClient});
   });
 
   it('handles "open"', async () => {
@@ -64,11 +62,8 @@ describe('handlers', () => {
       ),
     ).to.be.an.instanceof(Error);
 
-    handlers = handlers = makeHandlers({
-      makeVisualGridClient: () => ({
-        openEyes: openEyesWithCloseRejection,
-      }),
-    });
+    const visualGridClient = {openEyes: openEyesWithCloseRejection, closeBatch: () => {}};
+    handlers = makeHandlers({visualGridClient});
     handlers.batchStart({});
     expect(
       await handlers.checkWindow({}).then(
@@ -101,11 +96,8 @@ describe('handlers', () => {
       ),
     ).to.be.an.instanceof(Error);
 
-    handlers = makeHandlers({
-      makeVisualGridClient: () => ({
-        openEyes: openEyesWithCloseRejection,
-      }),
-    });
+    const visualGridClient = {openEyes: openEyesWithCloseRejection, closeBatch: () => {}};
+    handlers = makeHandlers({visualGridClient});
     handlers.batchStart({});
     expect(
       await handlers.close().then(
@@ -394,20 +386,22 @@ describe('handlers', () => {
     });
   });
 
-  it('handles "close"', async () => {
+  it('handles "batchStart"', async () => {
+    runningTests.add({test: 'test'});
     handlers.batchStart({});
-    const {checkWindow, close} = await handlers.open({__test: 123});
-
-    expect((await checkWindow()).__test).to.equal('checkWindow_123');
-    expect((await close()).__test).to.equal('close_123');
+    expect(runningTests.tests).to.deep.equal([]);
   });
 
-  it('handles "batchStart"', () => {
+  it('handles "close"', async () => {
     let flag;
-    handlers = makeHandlers({
-      makeVisualGridClient: () => (flag = 'flag'),
-    });
+    const open = () => ({checkWindow: async x => x, close: () => (flag = 'flag')});
+    const visualGridClient = {
+      openEyes: open,
+    };
+    handlers = makeHandlers({visualGridClient});
     handlers.batchStart({});
+    await handlers.open({accessibilityValidation: ''});
+    await handlers.close();
     expect(flag).to.equal('flag');
   });
 
@@ -423,9 +417,9 @@ describe('handlers', () => {
 
       abort: async () => {},
     });
-
+    const visualGridClient = {openEyes};
     handlers = makeHandlers({
-      makeVisualGridClient: () => ({openEyes}),
+      visualGridClient,
       processCloseAndAbort: async ({runningTests}) =>
         Promise.all(
           runningTests.map(async ({closePromise, abort}) =>
@@ -530,13 +524,12 @@ describe('handlers', () => {
   });
 
   it('error in openEyes should cause close to do nothing', async () => {
-    handlers = makeHandlers({
-      makeVisualGridClient: () => ({
-        openEyes: async () => {
-          throw new Error('open');
-        },
-      }),
-    });
+    const visualGridClient = {
+      openEyes: async () => {
+        throw new Error('open');
+      },
+    };
+    handlers = makeHandlers({visualGridClient});
     handlers.batchStart({});
     await handlers.open({}).catch(x => x);
     const err = await handlers.close().then(