+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- use global cypress hooks with `experimentalRunEvents` flag
 
 ## 3.19.1 - 2021/2/24